feat: :sparkles: Get min and max melting points

+++ app/src/main/java/nz/ac/aucklanduni/se306project1/models/items/ChemmatItem.java
@@ -3,17 +3,17 @@ package nz.ac.aucklanduni.se306project1.models.items;
 import java.util.List;
 
 public class ChemmatItem extends Item {
-    private double meltingPoint;
+    private float meltingPoint;
 
     public ChemmatItem() {
     }
 
-    public ChemmatItem(final String id, final String displayName, final String categoryId, final String description, final double price, final List<ColouredItemInformation> colours, final double meltingPoint) {
+    public ChemmatItem(final String id, final String displayName, final String categoryId, final String description, final double price, final List<ColouredItemInformation> colours, final float meltingPoint) {
         super(id, displayName, categoryId, description, price, colours);
         this.meltingPoint = meltingPoint;
     }
 
-    public double getMeltingPoint() {
+    public float getMeltingPoint() {
         return this.meltingPoint;
     }
 }
